pyproject.toml
[project]
name = "telegram-spam-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "telethon>=1.40.0",
]
